third_party/blink/web_tests/external/wpt/fetch/fetch-later/send-on-discard.tentative.https.window-expected.txt

This is a testharness.js-based test.
FAIL A discarded document sends all its fetchLater requests.  assert_equals: Number of sent beacons does not match expected count: expected 40 but got 0
FAIL A discarded document does not send an already aborted fetchLater request.  assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
Harness: the test ran to completion.
